requirements.txt
Pillow>=10.0.0
imagehash>=4.3.1
opencv-python>=4.8.0
tqdm>=4.66.0
numpy>=1.24.0
